PP_APP/requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
astroid==2.4.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
Django==3.0.5
django-cleanup==4.0.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
gast==0.3.3
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-vision==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
imageio==2.8.0
isort==4.3.21
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
matplotlib==3.2.1
mccabe==0.6.1
networkx==2.4
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==7.1.1
pkg-resources==0.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-resize-image==1.1.19
pytz==2019.3
PyWavelets==1.1.1
requests==2.23.0
rsa==4.0
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-serving-api==1.14.0
termcolor==1.1.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
<<<<<<< HEAD
opencv-python==4.2.0.34
tensorflow-serving-api==2.1.0
scikit-image==0.16.2
python-resize-image==1.1.19
django-cleanup==4.0.0

=======
Werkzeug==1.0.1
wrapt==1.12.1
>>>>>>> 68370ef... - Reorganized models and improved their naming schemes.
